+++ client/src/component/index.tsx
@@ -3,8 +3,8 @@ import FlipContainer from './layout/FlipContainer';
 import DragContainer from './layout/DragContainer';
 import ModalContainer from './modal/ModalContainer';
 import NavContainer from './layout/NavContainer';
-import HeaderButton from './button/HeaderButton';
-import OptionButton from './button/OptionButton';
+import TransperentButton from './button/TransperentButton';
+import OutlineButton from './button/OutlineButton';
 import CircleButton from './button/CircleButton';
 import Img from './img/Img';
 
@@ -14,8 +14,8 @@ export {
   FlipContainer,
   DragContainer,
   ModalContainer,
-  HeaderButton,
-  OptionButton,
+  TransperentButton,
+  OutlineButton,
   CircleButton,
   Img,
 };
