+++ client/src/page/MainPage/Header/CategoryTab/index.tsx
@@ -1,4 +1,4 @@
-import { HeaderButton } from '../../../../component';
+import { TransperentButton } from '../../../../component';
 import { click } from '../../../../util/pointerEvent';
 import styles from './CategoryTab.module.scss';
 
@@ -13,13 +13,14 @@ const CategoryTab = ({ id, name, activeCategoryId, onClick }: props) => {
   const changeActiveCategory = click(10, onClick, id);
 
   return (
-    <HeaderButton
+    <TransperentButton
       key={id}
       onPointerDown={changeActiveCategory}
-      value={name}
       isSelected={id === activeCategoryId}
       className={styles.button}
-    />
+    >
+      <div>{name}</div>
+    </TransperentButton>
   );
 };
